fix(trace viewer): splitting networkTab & networkResourceDetails components into separate files.
