+++ src/cli/traceViewer/web/ui/networkTab.css
@@ -24,100 +24,3 @@
 .network-tab:focus {
   outline: none;
 }
-
-.network-request {
-  box-shadow: var(--box-shadow);
-  white-space: nowrap;
-  display: flex;
-  align-items: center;
-  padding: 0 10px;
-  margin-bottom: 10px;
-  background: #fdfcfc;
-  width: 100%;
-  border: 3px solid transparent;
-  flex: none;
-  outline: none;
-}
-
-.network-request-title {
-  height: 36px;
-  display: flex;
-  align-items: center;
-  flex: 1;
-}
-
-.network-request-title-status {
-  font-weight: bold;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  padding: 0px 5px;
-  margin-right: 5px;
-}
-
-.status-success {
-  background-color: rgb(124, 187, 62);
-}
-
-.status-failure {
-  background-color: rgb(211, 35, 35);
-  color: white;
-}
-
-.status-neutral {
-  background-color: white;
-}
-
-.network-request-title-method {
-  font-weight: bold;
-}
-
-.network-request-title-url {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  flex: 1;
-}
-
-.network-request-title-content-type {
-  font-weight: bold;
-}
-
-.network-request-details {
-  font-family: var(--monospace-font);
-  width: 100%;
-}
-
-.network-request-details-url {
-  white-space: normal;
-  word-wrap: break-word;
-}
-
-.network-request-headers {
-  white-space: pre;
-  overflow: hidden;
-}
-
-.network-request-body {
-  white-space: pre;
-  overflow: scroll;
-  background-color: #dcdcdb;
-  border: black 1px solid;
-  max-height: 500px;
-}
-
-.network-request-response-body {
-  white-space: pre;
-  overflow: scroll;
-  background-color: #dcdcdb;
-  border: black 1px solid;
-  max-height: 500px;
-}
-
-.network-request.selected,
-.network-request:hover {
-  border-color: var(--inactive-focus-ring);
-}
-
-.network-request.selected:focus {
-  border-color: var(--orange);
-}
